requirements.txt
Flask
requests
python-dotenv
flask_login
flask_sqlalchemy
psycopg2
werkzeug

